+++ cli/Cargo.toml
@@ -80,7 +80,6 @@ libc = "=0.2.126"
 log = { version = "=0.4.17", features = ["serde"] }
 mitata = "=0.0.7"
 monch = "=0.2.0"
-node_resolver = "=0.1.1"
 notify = "=5.0.0-pre.15"
 once_cell = "=1.12.0"
 os_pipe = "=1.0.1"
